Removed a message about Liscense Violation and added more credits
I am following the rules, and Idk how it checks for Liscense violation, so I removed the message. On the bottom of every page on my site, have I added a link to this fork of Project-Lighthouse aswell as a Credits button, with everyone involved with Project-Lighthouse (That is listed as contributer on their Github)

I added three more names to the credits.