Changed some text and styles
Changed the style of the Register page to be the same as the login page, and added additional text to warn users that the username has to match up with their PSN ID or RPCN ID perfectly. Also changed the text on the landing page where it mentions you had to contact an admin to have an account made.